chat_bot/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.2
pydantic-settings>=2.0.3
python-dotenv>=1.0.0
langchain>=0.0.325
langchain-google-genai>=0.0.3
pytest>=7.4.3
httpx>=0.25.0
python-multipart>=0.0.6

